# Improve perf of LocalDate.WithCalendar if calendar is unchanged