--- a/src/NodaTime/LocalDate.cs
+++ b/src/NodaTime/LocalDate.cs
@@ -627,7 +627,9 @@ namespace NodaTime
         public LocalDate WithCalendar(CalendarSystem calendar)
         {
             Preconditions.CheckNotNull(calendar, nameof(calendar));
-            return new LocalDate(DaysSinceEpoch, calendar);
+            return ReferenceEquals(calendar, Calendar)
+                ? this
+                : new LocalDate(DaysSinceEpoch, calendar);
         }
 
         /// <summary>