If the LocalDate.WithCalendar is called with the same calendar the cost of the DaysSinceEpoch math can be avoided.

BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3880/23H2/2023Update/SunValley3) 12th Gen Intel Core i7-1270P, 1 CPU, 16 logical and 12 physical cores .NET SDK 8.0.303
  [Host]     : .NET 8.0.7 (8.0.724.31311), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.7 (8.0.724.31311), X64 RyuJIT AVX2

| Method                          | Mean     | Error     | StdDev    | Ratio |
|-------------------------------- |---------:|----------:|----------:|------:|
| WithCalendar_Unchanged_New      | 1.045 ns | 0.0150 ns | 0.0133 ns |  0.15 |
| WithCalendar_Unchanged_Previous | 7.043 ns | 0.0941 ns | 0.0880 ns |  1.00 |